Fixed scss-lint errors in _masthead.scss and video_player.scss
