+++ src/components/masthead/masthead.hbs
@@ -97,7 +97,7 @@
 
 </article>
 
-<div id="masthead-video-overlay" class="video-overlay">
+<div id="video-overlay" class="video-overlay">
   <div class="video-overlay__close">
     <a class="video-overlay__close__button icon-close-small js-close-video"></a>
   </div>
