+++ src/components/masthead/masthead_component.js
@@ -1,7 +1,6 @@
 import { Component } from "../../core/bane";
 import Slideshow from "../slideshow";
 import assign from "lodash/object/assign";
-// import Overlay from "../overlay";
 import "./masthead_nav.js";
 import coverVid from "../../core/utils/covervid";
 import MobilUtil from "../../core/mobile_util";
@@ -21,8 +20,6 @@ export default class MastheadComponent extends Component {
     this.events = {
       "click .js-play-video": "playVideo"
     };
-
-    // this.overlay = new Overlay();
 
     $.each(this.$straplines, function(index, strapline) {
       if (!$(strapline).html()) {
@@ -55,8 +52,9 @@ export default class MastheadComponent extends Component {
       minFontSize: 56
     });
 
+    // Initialize video overlay
     Video
-      .addPlayer("#masthead-video-overlay", "brightcove_default")
+      .addPlayer("#video-overlay", "brightcove")
       .then(this.playerReady.bind(this));
 
     this.subscribe();
@@ -78,49 +76,33 @@ export default class MastheadComponent extends Component {
    * Play the video, callback from click handler
    */
   playVideo() {
-    // this.overlay.show();
     this.player.play();
   }
 
   /**
-   * Callback from the player load event
+   * Callback from the player "ready" event
    * @param  {VideoPlayer} player Instance of the VideoPlayer
-   * @listens {play}
    */
   playerReady(player) {
     this.player = player;
-
-    // this.player.search(window.lp.place.atlasId)
-    //   .then(this.searchDone.bind(this));
     this.player.search().then(this.searchDone.bind(this));
-
-    // this.overlay.show();
-
-    // this.listenTo(this.player, "play", this.onPlay);
-    // this.listenTo(this.player, "stop", this.onStop);
-    // this.listenTo(this.player, "pause", this.onStop);
   }
 
-  // onPlay() {
-  //   // Use this?
-  // }
-
-  // onStop() {
-  //   // Use?
-  //   // this.overlay.hide();
-  // }
-
+  /**
+   * Callback from the player search()
+   * @param  {videos} list of video ids that matched the search
+   */
   searchDone(videos) {
     if (videos.length) {
-      // this.$el.find(".js-play-video")
-      //   .removeAttr("hidden")
-      //   .addClass("is-visible");
-
       let videoId = videos[0];
       this.player.loadVideo(videoId).then(this.loadDone.bind(this));
     }
   }
 
+  /**
+   * Callback from the player loadVideo()
+   * @param  {success} bool depicting whether the video successfully loaded or not
+   */
   loadDone(success) {
     if (!success) {
       return;
