--- src/components/video/brightcove.hbs
@@ -1,25 +1,0 @@
-<div class="video-overlay__close">
-  <a class="video-overlay__close__button icon-close-small js-close-video"></a>
-</div>
-<div class="masthead-video">
-  <div class="masthead-video__container">
-    <object id="masthead-video-player" class="BrightcoveExperience" style="z-index: -20;">
-      <param name="bgcolor" value="#FFFFFF" />
-      <param name="width" value="1290" />
-      <param name="height" value="725" />
-      <param name="playerID" value="4345531714001" />
-      <param name="playerKey" value="AQ~~,AAACtzWg-kk~,J9tdyvsy-9fKw2FQfVMUZOfR24XQMJiE" />
-      <param name="isVid" value="true" />
-      <param name="isUI" value="true" />
-      <param name="dynamicStreaming" value="true" />
-      <param name="htmlFallback" value="true" />
-      <param name="templateLoadHandler" value="BCL.player{{playerId}}.onTemplateLoad" />
-      <param name="templateReadyHandler" value="BCL.player{{playerId}}.onTemplateReady" />
-      <param name="includeAPI" value="true" />
-      <param name="@videoPlayer" value="4344778268001" />
-      <param name="forceHTML" value="true" />
-      <param name="secureConnections" value="true" />
-      <param name="secureHTMLConnections" value="true" />
-    </object>
-  </div>
-</div>
